Updated with mine and Matt's recent development

+++ holidayHomes_build/roomorama/Tables/roomorama.imp_amenity.sql
@@ -2,7 +2,7 @@
     [sourceId]     INT              NOT NULL,
     [runId]        INT              NOT NULL,
     [fileId]       INT              NOT NULL,
-    [id]           DECIMAL (28, 10) NULL,
+    [id]           INT NULL,
     [amenityValue] NVARCHAR (255)   NULL
 );
